Add Tag as abstract class and create new Group and Module Tag
