+++ src/main/java/seedu/address/model/person/TelegramHandle.java
@@ -1,0 +1,2 @@
+package seedu.address.model.person;public class TelegramHandle {
+}
