+++ src/main/java/seedu/address/model/person/User.java
@@ -1,0 +1,2 @@
+package seedu.address.model.person;public class User {
+}
